Move `default_logger` to a thread local
In the new design every thread gets its own logger (with a level,
pattern, and sink). We want all loggers to share the same sink though -
to do that we also expose setting the default sink used by newly minted
loggers. This won't change the value of existing loggers, but will
persist the sink used when creating new loggers.